fix bag save level

+++ Assets/UI/Scripts/StartMenu.cs
@@ -19,7 +19,7 @@ public class StartMenu : MonoBehaviour
 
     private void OnEnable()
     {
-        //InterstitialAd.Show();
+        InterstitialAd.Show();
         _startButton.onClick.AddListener(OnPlayButtonClick);
         _continueButton.onClick.AddListener(OnContinueButtonClick);
         _selectlanguageButton.onClick.AddListener(OnSelectLanguageButtonClick);
@@ -39,9 +39,13 @@ public class StartMenu : MonoBehaviour
 
     private void OnPlayButtonClick()
     {
-        PlayerPrefs.DeleteKey(_score);
         PlayerPrefs.SetInt(_scenes, _currentOpenLevels);
         PlayerPrefs.SetInt(Mute, 0);
+
+        if (_currentOpenLevels != 0)
+        {
+            PlayerPrefs.DeleteKey(_score);
+        }
 
         LoadMainMenu();
     }
